make it look nicer

+++ AppServer/models/Timezone.js
@@ -1,23 +1,25 @@
 var db=require('../dbconnection');
 
 var Timezone={
-	getAllTimezones:function(callback){
-		return db.query("Select id,name,city,gmt_offset from timezones",callback);
+	getAllTimezones:function(User,callback){
+		let where_clause = ((User.role_name === 'user')||(User.role_name === 'manager')) ? "where owner_id = " + User.id : "";
+		console.log(where_clause);
+		return db.query("Select id,name,city,gmt_offset from timezones " + where_clause,callback);
 	},
-	getTimezoneById:function(id,callback){
-		console.log(id);
-    return db.query("select id,name,city,gmt_offset from timezones where id=?",[id],callback);
+	getTimezoneById:function(User,id,callback){
+		let owner_id_clause = ((User.role_name === 'user')||(User.role_name === 'manager')) ? "and owner_id = " + User.id : "";
+    return db.query("select id,name,city,gmt_offset from timezones where id=? " + owner_id_clause,[id],callback);
   },
-  addTimezone:function(Timezone,callback){
-  	console.log("in addTimezone");
-  	console.log(Timezone);
-  	return db.query("Insert into timezones (name,city,gmt_offset)values(?,?,?)",[Timezone.name,Timezone.city,Timezone.gmt_offset],callback);
+  addTimezone:function(User, Timezone,callback){
+  	return db.query("Insert into timezones (name,city,gmt_offset,owner_id)values(?,?,?,?)",[Timezone.name,Timezone.city,Timezone.gmt_offset,User.id],callback);
   },
-  deleteTimezone:function(id,callback){
-    return db.query("delete from timezones where id=?",[id],callback);
+  deleteTimezone:function(User,id,callback){
+  	let owner_id_clause = ((User.role_name === 'user')||(User.role_name === 'manager')) ? "and owner_id = " + User.id : "";
+    return db.query("delete from timezones where id=? " + owner_id_clause,[id],callback);
 	},
-	updateTimezone:function(id,Timezone,callback){
-    return db.query("update timezones set name=?,city=?,gmt_offset=? where id=?",[Timezone.name,Timezone.city,Timezone.gmt_offset,id],callback);
+	updateTimezone:function(User,id,Timezone,callback){
+		let owner_id_clause = ((User.role_name === 'user')||(User.role_name === 'manager')) ? "and owner_id = " + User.id : "";
+    return db.query("update timezones set name=?,city=?,gmt_offset=? where id=? " + owner_id_clause,[Timezone.name,Timezone.city,Timezone.gmt_offset,id],callback);
 	},
 	deleteAll:function(item,callback){
 		var delarr=[];
